attempt to add error alert to failed query

+++ functions/chore_fxn.php
@@ -4,18 +4,21 @@
     function display_rows(){
         $var_data = all_chores();
 
-        foreach ($var_data as $row){
-            echo "<div class='row'>
-            <div>".$row['chorename']."</div>
-                <div class='actions'>
-                    <div><span class='material-symbols-outlined'>edit</span></div>
-                    <div>
-                        <a style='color:rgb(241, 143, 143);' href='../action/delete_chore_action.php?id=".$row['cid']."'>
-                            <span class='material-symbols-outlined'>delete</span>
-                        </a>
+        if (!empty($var_data)){
+
+            foreach ($var_data as $row){
+                echo "<div class='row'>
+                <div>".$row['chorename']."</div>
+                    <div class='actions'>
+                        <div><span class='material-symbols-outlined'>edit</span></div>
+                        <div>
+                            <a style='color:rgb(241, 143, 143);' href='../action/delete_chore_action.php?id=".$row['cid']."'>
+                                <span class='material-symbols-outlined'>delete</span>
+                            </a>
+                        </div>
                     </div>
-                </div>
-            </div>";
+                </div>";
+            }
         }
     }
